+++ client/src/components/community/MarketViewForm.js
@@ -147,11 +147,19 @@ const MarketViewForm = ({ postType }) => {
   };
 
   const handleRemoveButton = async () => {
+    if (!window.confirm("게시글을 삭제하시겠습니까?")) {
+      alert("취소 했습니다..");
+      return;
+    }
+
     try {
       await axios.delete(port.url + `/api/market/list/${paramsId}/delete`, {
         headers: { accessToken: cookies.userData.accessToken },
       });
     } catch {}
+
+    alert("게시글을 삭제했습니다.");
+
     navigate("/market");
     window.location.reload();
   };
