+++ client/src/components/community/OotdUpdateForm.js
@@ -60,6 +60,11 @@ const OotdUpdateForm = ({ postType, images }) => {
 
   //글 삭제하기 버튼 클릭시 이벤트
   const handleRemoveButton = async () => {
+    if (!window.confirm("게시글을 삭제하시겠습니까?")) {
+      alert("취소 했습니다.");
+      return;
+    }
+
     try {
       await axios.delete(port.url + `/api/posts/list/${paramsId}/delete`, {
         headers: { accessToken: cookies.userData.accessToken },
@@ -67,7 +72,9 @@ const OotdUpdateForm = ({ postType, images }) => {
     } catch (e) {
       console.log(e);
     }
+    alert("게시글을 삭제했습니다.");
     navigate("/board");
+    window.location.reload();
   };
 
   const handleChange = (e) => {
